+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -92,5 +92,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-01T17:40:07.626Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T05:22:47.672Z")
     r"""The date and time the resource was created"""