+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQS70Y0B4TXQJK29Q28XD6BZ"
+        "01JQTF7J33DXRN48PFBWHWD430"
     )
     r"""The id of the resource"""