+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQS70Y0CET2HTM450GH980EZ"
+        "01JQTF7J340MXQTVE096BNW3V6"
     )
     r"""The id of the resource"""